LA Changes, Other Fixes
Ported some changes from the surprising new DS beta update, fixed the lights for the ruins light rays (Maybe incidentally fixed for the rainforest, too?), and made the Living Artifact's music work with the new seamless swapping system.
Also finally added the LA character to the seamless swap list, so they no longer show up in the character selection screen and crash the game!

Also tried to make the screenshake work again during the equip animation, but no dice so far.
I did hack at the LA's stategraph a bit, so now it can sort-of fire, but for whatever reason the large bolt doesn't happen, so the best I've got is firing the little one.
I also moved the shoot code from the stategraph itself to the character, but that's probably not doing anything, since it was likely working the same when it was in there. May or may not revert that at some point since it seem to just be needless deviation from Hamlet.