Added the Smelter + Bugfixes
The smelter has been added
Added alloy alongside it
I removed the duplicate rocks.lua and made one with only the Hamlet-exclusive stuff, so the original DST rocks work again
I fixed the city lamp's light/unlight function, so it properly enables at night/dusk and disables in the day
Disabled the recipe for the really old testing house, to prevent more confusion about which one to build
Added the reconstruction project, so destroyed buildings no longer crash, and can be rebuilt with a hammer
Added recipes for a bunch of stuff. The tabs are still sort-of placeholder, because we need the key and new city & spelunking tabs to be added, but it's a start!