Rift Watch Portal Works
Wanda's Rift Watch portal now works through the interior system! So you can send your friends or heavy things in to or out of interior space, or through different rooms, with ease!

Also a bunch of fixes to use cases for the watches.

I don't _really_ like that I'm copying so much of the original files to make this work, but it's the only way I can really think of without some hacky nonsense.